Update of Maintainer Email Address and Initial Makefile Installation
Updated the email address of the maintainer since he has changed institutions.
Added Makefiles in the runtime and tools subdirectories to those that the
configure script initially copies to the object tree.  This eases updates to
newer versions of LLVM as the object tree can be removed and reconfigured
easily.



git-svn-id: http://llvm.org/svn/llvm-project/safecode/trunk@247515 91177308-0d34-0410-b5e6-96231b3b80d8